refactor(ShopPage): resolving comments on ShopPage and use checkbox
