--- src/components/ShopPageButton.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-interface ShopPageButton {
-  children: React.ReactNode;
-}
-const ShopPage_button: React.FC<ShopPageButton> = ({ children }) => {
-  return (
-    <button className="w-60 px-[65px] max-h-10 py-2 text-center bg-on-primary-light text-black  rounded-full text-base cursor-pointer">
-      {children}
-    </button>
-  );
-};
-
-export default ShopPage_button;
