+++ src/pages/ShopPage.tsx
@@ -1,6 +1,6 @@
 import Sidebar from "../components/ui/Sidebar";
-import ShopPage_aside from "../components/ShopPageAside";
-import ShopPage_Products from "../components/ShopPageProducts";
+import ShowPageAside from "../components/ShopPageAside";
+import ShopPageProducts from "../components/ShopPageProducts";
 const ShopPage = () => {
   return (
     <div className="flex  bg-background-base-light justify-between font-yekan-bakh">
@@ -11,9 +11,9 @@ const ShopPage = () => {
       {/* main + aside */}
       <div className="relative flex flex-row gap-16 top-8 left-[156px] justify-center font-YekanBakh text-black">
         {/* aside */}
-        <ShopPage_aside />
+        <ShowPageAside/>
         {/* products card */}
-        <ShopPage_Products />
+        <ShopPageProducts />
       </div>
     </div>
   );
